abc_requirement.txt
sqlite
absl-py==0.11.0
asgiref==3.3.1
astunparse==1.6.3
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
dj-database-url==0.5.0
Django==2.1.15
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-extensions==3.0.4
django-nested-admin==3.3.2
django-print-sql==2018.3.6
django-rest-framework==0.1.0
django-xhtml2pdf==0.0.4
djangorestframework==3.11.1
flatbuffers==1.12
gast==0.3.3
google-auth==1.25.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imantics==0.1.12
imgaug==0.4.0
importlib-metadata==3.4.0
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
labelme2coco==0.1.2
lxml==4.6.2
Markdown==3.3.3
matplotlib==3.2.0
networkx==2.5
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
Pillow==8.1.0
pixellib==0.5.2
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-image==0.18.1
scipy==1.6.0
Shapely==1.7.1
six==1.15.0
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tifffile==2021.2.1
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
xmljson==0.2.1
zipp==3.4.0
